+++ packages/soda-hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # soda-hooks
 
+## 6.14.4
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.42.0
+    - soda-type@6.5.0
+
 ## 6.14.3
 
 ### Patch Changes