+++ packages/soda-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-react
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.42.0
+
 ## 0.0.26
 
 ### Patch Changes